// data/sample-data-type.ts
export type sampleDataType = {
  kidney: {
    urine_ph: 6
    uric_acid: 7.8
    bpa: null
    egfr: null
    heavy_metals: {
      mercury: null
      lead: null
      cadmium: null
    }
    bun: 6
    hydration: null
    random_urine_albumin: null
  }
  heart: {
    homocysteine: null
    heart_rate: null
    imt: null
    nt_bnp: null
    heart_recovery_rate: null
    cholesterol: 194
    hs_crp: 12.95
    troponin: null
    heart_rate_variability: null
    blood_pressure: {
      systolic: null
      diastolic: null
    }
    aortic_compliance: null
  }
  liver: {
    ggt: 37
    ast: 26
    bilirubin: 0.44
    alt: 29
  }
  lungs: {
    pulmonary_capacity: null
    chest_impression_log: ''
    vo2_max: null
    heart_rate_variability: null
    spo2: null
  }
  brain: {
    rmssd: null
    p_tau217: null
    neurotransmitters: {
      serotonin: null
      gaba: null
      dopamine: null
    }
    gait_readiness: null
    sdnn: null
    brain_impression_log: ''
    cd_risc: null
  }
  hormonal_reproductive: {
    women: null
    men: {
      insulin: null
      fsh: null
      cmp: null
      estrone: null
      ft4: null
      psa: null
      dhea_s: null
      hba1c: null
      prolactin: null
      hs_crp: null
      tsh: null
      dht: null
      ft3: null
      vitamin_d_25oh: 7.9
      homocysteine: null
      parathyroid_hormone: null
      androstenedione: null
      testosterone_total: null
      estradiol: null
      igf_1: null
      lipids: null
      progesterone: null
      testosterone_free: null
      lh: null
      shbg: null
    }
    general: {
      hba1c: 5.3
      hormones_panel: ''
      igf_1: null
    }
  }
}